+++ path/dasher.go
@@ -1,7 +1,7 @@
 // Copyright 2010 The draw2d Authors. All rights reserved.
 // created: 13/12/2010 by Laurent Le Goff
 
-package draw2d
+package path
 
 type DashVertexConverter struct {
 	next           LineBuilder
